+++ packages/@coorpacademy-app-review/public/sandbox.tsx
@@ -4,7 +4,7 @@ import isNil from 'lodash/fp/isNil';
 import get from 'lodash/fp/get';
 import pipe from 'lodash/fp/pipe';
 
-import WebContext from '@coorpacademy/components/es/atom/web-context';
+import {WebContext} from '@coorpacademy/components/es/atom/provider';
 import AppReview from '../src';
 import {services} from '../src/test/util/services.mock';
 import type {AppOptions} from '../src/types/common';
