+++ packages/@coorpacademy-components/src/atom/provider/index.js
@@ -2,6 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import {defaultsDeep} from 'lodash/fp';
 import {SrcPropType, ColorPropType, HexPropType} from '../../util/proptypes';
+import WebContext, {useWebContext} from './web-context';
 
 const DEFAULT_SKIN = {
   common: {
@@ -98,4 +99,5 @@ class Provider extends React.Component {
   }
 }
 
+export {WebContext, useWebContext};
 export default Provider;
