--- packages/@coorpacademy-components/src/atom/web-context/test/fixtures/default.js
@@ -1,13 +1,0 @@
-import React from 'react';
-
-export default {
-  props: {
-    skin: {
-      common: {
-        primary: '#00B0FF'
-      }
-    },
-    translate: v => v
-  },
-  children: <div />
-};
